+++ src/main/java/cz/jakubfajkus/reservations/utils/ReservationPriceCalculator.java
@@ -2,6 +2,8 @@ package cz.jakubfajkus.reservations.utils;
 
 import cz.jakubfajkus.reservations.dto.ReservationDTO;
 
+import java.math.BigDecimal;
+
 public interface ReservationPriceCalculator {
-    long calculate(ReservationDTO reservation);
+    BigDecimal calculate(ReservationDTO reservation);
 }
